pyproject.toml
[project]
name = "sync-clip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=10.0.0",
    "pyperclip>=1.8.0",
    "psutil>=5.9.0",
]
